requirements.txt
alembic==1.8.0
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
